w3c-reports/12092014/gtk/runned-tests/XMLHttpRequest/send-authentication-competing-names-passwords-expected.txt


PASS XMLHttpRequest: send() - "Basic" authenticated requests with competing user name/password options user/pass in open() call 
PASS XMLHttpRequest: send() - "Basic" authenticated requests with competing user name/password options another user/pass in open() call - must override cached credentials from previous test 
PASS XMLHttpRequest: send() - "Basic" authenticated requests with competing user name/password options user/pass both in URL userinfo AND open() call - expexted that open() wins 
FAIL XMLHttpRequest: send() - "Basic" authenticated requests with competing user name/password options user/pass *only* in URL userinfo assert_equals: expected "6bf1b349-5e09-4eed-a2f6-56954dc92e27\nbe46871a-bcd1-4df2-8102-f20caea3bf89" but got "78159ef5-5e3b-4b19-9484-5757d88c2496\naf64a345-d8b0-46bd-8e97-b21b8695eccd"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/XMLHttpRequest/send-authentication-competing-names-passwords.htm:37:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
request@http://localhost:8800/XMLHttpRequest/send-authentication-competing-names-passwords.htm:16:13
global code@http://localhost:8800/XMLHttpRequest/send-authentication-competing-names-passwords.htm:49:14)
FAIL XMLHttpRequest: send() - "Basic" authenticated requests with competing user name/password options user name in URL userinfo, password in open() call: user name wins and password is thrown away assert_equals: expected "e7426e8d-2f80-4cb0-93b6-1847037bcdd5\n" but got "\n195f35be-3d9d-49b7-818b-30ad6db8aac8"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/XMLHttpRequest/send-authentication-competing-names-passwords.htm:37:24
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
request@http://localhost:8800/XMLHttpRequest/send-authentication-competing-names-passwords.htm:16:13
global code@http://localhost:8800/XMLHttpRequest/send-authentication-competing-names-passwords.htm:50:14)
PASS XMLHttpRequest: send() - "Basic" authenticated requests with competing user name/password options user name and password in URL userinfo, only user name in open() call: user name in open() wins 
